Improve screener logic
