+++ screeners/leader_stock/screener.py
@@ -97,6 +97,8 @@ class LeaderStockScreener:
         self.pe_map = {}
         self.revenue_growth_map = {}
         self.stock_rs_percentile = {}
+        self.market_cap_map = {}
+        self.ipo_date_map = {}
 
         # 섹터, P/E, 매출 성장률 메타데이터
         if os.path.exists(STOCK_METADATA_PATH):
@@ -108,6 +110,11 @@ class LeaderStockScreener:
                     self.pe_map = meta.set_index('symbol')['pe_ratio'].to_dict()
                 if 'revenue_growth' in meta.columns:
                     self.revenue_growth_map = meta.set_index('symbol')['revenue_growth'].to_dict()
+                if 'market_cap' in meta.columns:
+                    self.market_cap_map = meta.set_index('symbol')['market_cap'].to_dict()
+                if 'ipo_date' in meta.columns:
+                    meta['ipo_date'] = pd.to_datetime(meta['ipo_date'], errors='coerce')
+                    self.ipo_date_map = meta.set_index('symbol')['ipo_date'].to_dict()
             except Exception as e:
                 logger.warning(f"메타데이터 로드 실패: {e}")
         else:
@@ -152,6 +159,25 @@ class LeaderStockScreener:
     def _growth_slowdown(self, ticker):
         growth = self.revenue_growth_map.get(ticker)
         return growth is not None and growth < 15
+
+    def _small_cap_or_recent_ipo(self, ticker):
+        """소형주이거나 최근 IPO 여부 확인"""
+        cap = self.market_cap_map.get(ticker)
+        if cap is not None and cap < 2_000_000_000:
+            return True
+        ipo_date = self.ipo_date_map.get(ticker)
+        if ipo_date is not None:
+            return (self.today - ipo_date).days <= 5 * 365
+        return False
+
+    def _calculate_momentum_indicator(self, df):
+        """간단한 모멘텀 지표 (RSI + 5일 수익률)"""
+        if len(df) < 6:
+            return 0.0
+        rsi = df['rsi_14'].iloc[-1]
+        roc5 = (df['close'].iloc[-1] / df['close'].iloc[-5] - 1) * 100
+        score = rsi + roc5
+        return max(min(score, 100), 0)
     
     def calculate_sector_rs(self, sector_etfs):
         """섹터별 상대 강도(RS) 계산"""
@@ -395,10 +421,10 @@ class LeaderStockScreener:
             },
             "stage3": {
                 "market_trend": lambda df, recent, sector, strong_sectors, ticker=None: self._market_trend_ok(),
-                "small_cap_or_ipo": lambda df, recent, sector, strong_sectors, ticker=None: True,
+                "small_cap_or_ipo": lambda df, recent, sector, strong_sectors, ticker=None: self._small_cap_or_recent_ipo(ticker),
                 "volume_explosion": lambda df, recent, sector, strong_sectors, ticker=None: recent['volume_ratio'] >= 5.0,
                 "rsi_overbought": lambda df, recent, sector, strong_sectors, ticker=None: recent['rsi_14'] >= 70,
-                "momentum": lambda df, recent, sector, strong_sectors, ticker=None: (recent['close'] / df.iloc[-5]['close'] - 1) * 100 >= 10,
+                "momentum": lambda df, recent, sector, strong_sectors, ticker=None: self._calculate_momentum_indicator(df) >= 90,
             },
             "stage4": {
                 "high_pe": lambda df, recent, sector, strong_sectors, ticker=None: self._is_high_pe(ticker),
@@ -417,3 +443,4 @@ def run_leader_stock_screening():
 
 if __name__ == "__main__":
     run_leader_stock_screening()
+
