+++ screeners/momentum_signals/screener.py
@@ -11,6 +11,7 @@ from config import (
     DATA_US_DIR,
     RESULTS_DIR,
     US_WITH_RS_PATH,
+    STOCK_METADATA_PATH,
     MOMENTUM_SIGNALS_CRITERIA,
     MOMENTUM_SIGNALS_RESULTS_DIR,
 )
@@ -33,6 +34,7 @@ class MomentumSignalsScreener:
         self.today = get_us_market_today()
         ensure_dir(MOMENTUM_SIGNALS_RESULTS_DIR)
         self.rs_scores = self._load_rs_scores()
+        self._load_metadata()
         self.strong_sectors = self._load_sector_strength()
     
     def _calculate_macd(self, df, fast=12, slow=26, signal=9):
@@ -161,6 +163,75 @@ class MomentumSignalsScreener:
         df['ad'] = (mfm * df['volume']).cumsum()
         return df
 
+    def _detect_cup_and_handle(self, df, window: int = 180) -> bool:
+        """단순 컵앤핸들 패턴 인식"""
+        try:
+            from scipy.signal import find_peaks
+
+            if len(df) < window:
+                return False
+
+            data = df.tail(window)
+            prices = data['close'].values
+            peaks, _ = find_peaks(prices)
+            troughs, _ = find_peaks(-prices)
+            if len(peaks) < 2 or len(troughs) == 0:
+                return False
+
+            left = peaks[0]
+            right = peaks[-1]
+            bottom = troughs[troughs > left]
+            bottom = bottom[bottom < right]
+            if len(bottom) == 0:
+                return False
+            bottom = bottom[prices[bottom].argmin()]
+
+            left_high = prices[left]
+            right_high = prices[right]
+            bottom_low = prices[bottom]
+            depth = min(left_high, right_high) - bottom_low
+            if depth <= 0:
+                return False
+
+            depth_pct = depth / min(left_high, right_high) * 100
+            handle_low = prices[right: ].min()
+            handle_pct = (right_high - handle_low) / depth * 100
+
+            return (
+                abs(left_high - right_high) / min(left_high, right_high) <= 0.1
+                and depth_pct >= 20
+                and handle_pct <= 30
+            )
+        except Exception as e:
+            logger.debug(f"Cup&Handle detection error: {e}")
+            return False
+
+    def _load_metadata(self):
+        """Load sector and RS percentile information."""
+        self.sector_map = {}
+        self.stock_rs_percentile = {}
+
+        if os.path.exists(STOCK_METADATA_PATH):
+            try:
+                meta = pd.read_csv(STOCK_METADATA_PATH)
+                if {'symbol', 'sector'}.issubset(meta.columns):
+                    self.sector_map = meta.set_index('symbol')['sector'].to_dict()
+            except Exception as e:
+                logger.warning(f"메타데이터 로드 실패: {e}")
+        else:
+            logger.warning(f"메타데이터 파일이 없습니다: {STOCK_METADATA_PATH}")
+
+        if os.path.exists(US_WITH_RS_PATH) and self.sector_map:
+            try:
+                rs_df = pd.read_csv(US_WITH_RS_PATH)
+                if {'symbol', 'rs_score'}.issubset(rs_df.columns):
+                    rs_df['sector'] = rs_df['symbol'].map(self.sector_map)
+                    rs_df.dropna(subset=['sector'], inplace=True)
+                    rs_df['sector_rank'] = rs_df.groupby('sector')['rs_score'].rank(pct=True) * 100
+                    self.stock_rs_percentile = rs_df.set_index('symbol')['sector_rank'].to_dict()
+            except Exception as e:
+                logger.warning(f"RS 메타데이터 로드 실패: {e}")
+
     def _load_rs_scores(self):
         """RS 점수 로드"""
         if os.path.exists(US_WITH_RS_PATH):
@@ -215,9 +286,13 @@ class MomentumSignalsScreener:
                 if rs_score is None or rs_score < 70:
                     continue
 
-                # 섹터 상대 강도 필터 (임의 할당)
-                sector = np.random.choice(self.all_sectors) if hasattr(self, 'all_sectors') else None
+                # 실제 섹터 조회 및 강한 섹터 필터
+                sector = self.sector_map.get(ticker)
+                if not sector:
+                    continue
                 if self.strong_sectors and sector not in self.strong_sectors:
+                    continue
+                if self.stock_rs_percentile.get(ticker, 0) < 60:
                     continue
                 
                 # 데이터 로드
@@ -306,10 +381,13 @@ class MomentumSignalsScreener:
                 # 12. 신고가 돌파 (52주 신고가 돌파)
                 year_high = df.iloc[-260:]['high'].max() if len(df) >= 260 else df['high'].max()
                 signals['new_high'] = recent['close'] > year_high * 0.98  # 52주 신고가의 98% 이상
-                
+
                 # 13. 가격 모멘텀 (5일간 5% 이상 상승)
                 five_day_return = (recent['close'] / df.iloc[-6]['close'] - 1) * 100
                 signals['price_momentum'] = five_day_return >= 5
+
+                # 13-1. 컵앤핸들 패턴 돌파 여부
+                signals['cup_handle_breakout'] = self._detect_cup_and_handle(df)
                 
                 # 14. 이격도 양호 (종가가 20일 이동평균선 대비 8% 이내)
                 price_to_ma = (recent['close'] / recent['sma_20'] - 1) * 100
@@ -325,8 +403,9 @@ class MomentumSignalsScreener:
 
                 # 17. Stage 2A 확인 (10주 SMA 상승, 30주 SMA 수평/상승)
                 signals['stage_2a'] = (
-                    recent['sma_50'] > df.iloc[-10]['sma_50'] and
-                    recent['sma_30w'] >= df.iloc[-10]['sma_30w']
+                    len(df) >= 50
+                    and recent['sma_50'] > df.iloc[-50]['sma_50']
+                    and recent['sma_30w'] >= df.iloc[-50]['sma_30w']
                 )
 
                 # 18. VWAP 돌파
